Split statement tests out
